Conflicts:

	build/Makefile.cinelerra
	build/Makefile.esound
	build/Makefile.freetype
	build/Makefile.tiff
	build/Makefile.toolame
	build/Makefile.uuid
	build/configure.mix2005
	cinelerra/Makefile
	cinelerra/aautomation.C
	cinelerra/aboutprefs.C
	cinelerra/audio1394.C
	cinelerra/audiodevice.C
	cinelerra/auto.h
	cinelerra/automation.C
	cinelerra/automation.h
	cinelerra/automation.inc
	cinelerra/awindow.C
	cinelerra/awindowgui.h
	cinelerra/browsebutton.C
	cinelerra/browsebutton.h
	cinelerra/builddate.h
	cinelerra/clipedit.C
	cinelerra/cwindowtool.C
	cinelerra/device1394input.C
	cinelerra/dvbtune.C
	cinelerra/editpanel.C
	cinelerra/edits.C
	cinelerra/file.C
	cinelerra/file.inc
	cinelerra/fileavi.C
	cinelerra/filebase.C
	cinelerra/filemov.C
	cinelerra/filempeg.C
	cinelerra/filempeg.h.bak2
	cinelerra/filevorbis.C
	cinelerra/filevorbis.h
	cinelerra/filexml.C
	cinelerra/formatpresets.C
	cinelerra/formattools.C
	cinelerra/gwindowgui.C
	cinelerra/gwindowgui.h
	cinelerra/interfaceprefs.C
	cinelerra/loadfile.C
	cinelerra/loadfile.h
	cinelerra/localsession.C
	cinelerra/localsession.h
	cinelerra/main.C
	cinelerra/mainmenu.C
	cinelerra/mainundo.C
	cinelerra/mainundo.h
	cinelerra/maskauto.C
	cinelerra/maskautos.h
	cinelerra/mwindow.inc
	cinelerra/mwindowedit.C
	cinelerra/mwindowmove.C
	cinelerra/recordmonitor.C
	cinelerra/render.C
	cinelerra/render.h
	cinelerra/track.C
	cinelerra/trackcanvas.C
	cinelerra/trackcanvas.h
	cinelerra/tracks.h
	cinelerra/tracksedit.C
	cinelerra/transitionpopup.C
	cinelerra/undostack.C
	cinelerra/undostack.h
	cinelerra/undostack.inc
	cinelerra/vautomation.C
	cinelerra/viewmenu.C
	cinelerra/virtualvnode.C
	cinelerra/vmodule.C
	cinelerra/vrender.C
	cinelerra/vtracking.C
	cinelerra/zoombar.C
	configure
	doc/Makefile
	doc/cinelerra.html
	doc/cinelerra.texi
	global_config
	guicast/Makefile
	guicast/arraylist.h
	guicast/bctheme.C
	guicast/bctheme.h
	guicast/bctoggle.C
	guicast/bcwindowbase.C
	guicast/bcwindowbase.h
	guicast/bcwindowevents.C
	libavc1394-0.5.1/COPYING
	libmpeg3/Makefile
	libmpeg3/video/seek.c
	plugin_config
	plugins/1080to480/1080to480.C
	plugins/1080to480/1080to480.h
	plugins/1080to480/picon_png.h
	plugins/Makefile
	plugins/chromakey/chromakey.C.bak
	plugins/defaulttheme/defaulttheme.h
	plugins/deinterlace/deinterlace.C
	plugins/deinterlace/deinterwindow.C
	plugins/diffkey/diffkey.C.bak
	plugins/dissolve/dissolve.C.bak
	plugins/downsample/Makefile
	plugins/fonts/fonts.dir
	plugins/gradient/blur.inc
	plugins/histogram_bezier/histogram.C
	plugins/histogram_bezier/histogram.h
	plugins/histogram_bezier/histogram.inc
	plugins/histogram_bezier/histogramconfig.C
	plugins/histogram_bezier/histogramconfig.h
	plugins/histogram_bezier/histogramconfig.inc
	plugins/histogram_bezier/histogramwindow.C
	plugins/histogram_bezier/histogramwindow.h
	plugins/histogram_bezier/histogramwindow.inc
	plugins/histogram_bezier/max_picon_png.h
	plugins/histogram_bezier/mid_picon_png.h
	plugins/histogram_bezier/min_picon_png.h
	plugins/histogram_bezier/picon_png.h
	plugins/liveaudio/picon_png.h
	plugins/livevideo/Makefile
	plugins/motion/Makefile
	plugins/motion/affine.C
	plugins/motion/motion.C
	plugins/motion/motion.C.bak
	plugins/motion/motionwindow.C
	plugins/parametric/fourier.C
	plugins/parametric/fourier.h
	plugins/pitch/Makefile
	plugins/pitch/fourier.C
	plugins/pitch/fourier.h
	plugins/reframert/reframert.C
	plugins/rgb601/ivtc.C.2
	plugins/shapewipe/shapewipe.C
	plugins/shapewipe/shapewipe.h
	plugins/spectrogram/Makefile
	plugins/spectrogram/spectrogram.h
	plugins/suv/Makefile
	plugins/svg/svg.C
	plugins/threshold/threshold.C
	plugins/threshold/thresholdwindow.C
	plugins/threshold/thresholdwindow.h
	plugins/timestretch/timestretch.C
	plugins/timestretch/timestretch.h
	plugins/titler/Makefile
	plugins/titler/title.C
	plugins/titler/title.C.stroker
	plugins/titler/title.h
	plugins/titler/titlewindow.C
	plugins/titler/titlewindow.h
	plugins/titler/titlewindow.h.stroker
	plugins/videoscope/videoscope.C
	po/Makefile
	quicktime/Makefile
	quicktime/ffmpeg/COPYING
	quicktime/funcprotos.h
	quicktime/qtffmpeg.c
	quicktime/qth264.c
	quicktime/qtprivate.h
	quicktime/quicktime.c
	quicktime/yuv2.c
	thirdparty/libavc1394-0.5.1/COPYING

+++ cinelerra/render.C
@@ -1,8 +1,30 @@
+
+/*
+ * CINELERRA
+ * Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
+ * 
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * 
+ */
+
 #include "arender.h"
 #include "asset.h"
 #include "auto.h"
 #include "batchrender.h"
 #include "bcprogressbox.h"
+#include "bcsignals.h"
 #include "cache.h"
 #include "clip.h"
 #include "compresspopup.h"
@@ -311,65 +333,67 @@ void Render::stop_operation()
 void Render::run()
 {
 	int format_error;
-
+	const int debug = 0;
 
 	result = 0;
 
 	if(mode == Render::INTERACTIVE)
 	{
 // Fix the asset for rendering
-printf("Render::run 1\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 		Asset *asset = new Asset;
 		load_defaults(asset);
-printf("Render::run 2\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 		check_asset(mwindow->edl, *asset);
-printf("Render::run 3\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 
 // Get format from user
 		if(!result)
 		{
-printf("Render::run 4\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 			do
 			{
 				format_error = 0;
 				result = 0;
 
 				{
-printf("Render::run 5\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 					RenderWindow window(mwindow, this, asset);
-printf("Render::run 6\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 					window.create_objects();
-printf("Render::run 7\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 					result = window.run_window();
 printf("Render::run 8\n");
 					if (! result) {
 						// add to recentlist only on OK
 						window.format_tools->path_recent->add_item(FILE_FORMAT_PREFIX(asset->format), asset->path);
 					}
+if(debug) printf("Render::run %d\n%", __LINE__);
 				}
 
 				if(!result)
 				{
-printf("Render::run 8.1\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 // Check the asset format for errors.
 					FormatCheck format_check(asset);
-printf("Render::run 8.2\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 					format_error = format_check.check_format();
-printf("Render::run 8.3\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 				}
 			}while(format_error && !result);
 		}
-printf("Render::run 9\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 
 		save_defaults(asset);
 		mwindow->save_defaults();
-printf("Render::run 10\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 
 		if(!result) render(1, asset, mwindow->edl, strategy, range_type);
 printf("Render::run 11\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 
 		Garbage::delete_object(asset);
-printf("Render::run 12\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 	}
 	else
 	if(mode == Render::BATCH)
@@ -433,7 +457,7 @@ printf("Render::run 12\n");
 			mwindow->batch_render->update_done(-1, 0, 0);
 		}
 	}
-printf("Render::run 100\n");
+if(debug) printf("Render::run %d\n%", __LINE__);
 }
 
 
@@ -837,11 +861,12 @@ printf("Render::render 90\n");
 				printf("Render::render: Error rendering data\n");
 			}
 		}
+printf("Render::render 91\n");
 
 // Delete the progress box
 		stop_progress();
 
-//printf("Render::render 100\n");
+printf("Render::render 100\n");
 
 
 
@@ -858,6 +883,7 @@ printf("Render::render 90\n");
 	{
 		mwindow->gui->lock_window("Render::render 3");
 
+		mwindow->undo->update_undo_before();
 
 
 
@@ -876,7 +902,7 @@ printf("Render::render 90\n");
 
 
 		mwindow->save_backup();
-		mwindow->undo->update_undo(_("render"), LOAD_ALL);
+		mwindow->undo->update_undo_after(_("render"), LOAD_ALL);
 		mwindow->update_plugin_guis();
 		mwindow->gui->update(1, 
 			2,
@@ -889,6 +915,7 @@ printf("Render::render 90\n");
 		mwindow->gui->unlock_window();
 	}
 
+printf("Render::render 110\n");
 
 // Disable hourglass
 	if(mwindow)
@@ -911,7 +938,7 @@ printf("Render::render 90\n");
 	delete packages;
 	in_progress = 0;
 	completion->unlock();
-//printf("Render::render 120\n");
+printf("Render::render 120\n");
 
 	return result;
 }
@@ -1091,8 +1118,12 @@ RenderWindow::RenderWindow(MWindow *mwindow, Render *render, Asset *asset)
 
 RenderWindow::~RenderWindow()
 {
+//printf("RenderWindow::~RenderWindow %d\n", __LINE__);
 	delete format_tools;
+//sleep(1);
+//printf("RenderWindow::~RenderWindow %d\n", __LINE__);
 	delete loadmode;
+//printf("RenderWindow::~RenderWindow %d\n", __LINE__);
 }
 
 
@@ -1105,9 +1136,9 @@ int RenderWindow::load_profile(int profile_slot)
 
 
 
-int RenderWindow::create_objects()
+void RenderWindow::create_objects()
 {
-	int x = 5, y = 5;
+	int x = 10, y = 5;
 	add_subwindow(new BC_Title(x, 
 		y, 
 		(char*)((render->strategy == FILE_PER_LABEL || 
@@ -1164,7 +1195,6 @@ int RenderWindow::create_objects()
 	add_subwindow(new BC_OKButton(this));
 	add_subwindow(new BC_CancelButton(this));
 	show_window();
-	return 0;
 }
 
 void RenderWindow::update_range_type(int range_type)
